refactor: Reorganize notification package structure
