+++ android/app/src/main/java/chat/rocket/reactnative/MainApplication.kt
@@ -12,7 +12,6 @@ import com.facebook.react.ReactPackage
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint.load
 import com.facebook.react.defaults.DefaultReactHost.getDefaultReactHost
 import com.facebook.react.defaults.DefaultReactNativeHost
-import com.facebook.react.flipper.ReactNativeFlipper
 import com.facebook.soloader.SoLoader
 import com.nozbe.watermelondb.jsi.WatermelonDBJSIPackage;
 import com.facebook.react.bridge.JSIModulePackage;
@@ -24,6 +23,7 @@ import com.wix.reactnativenotifications.core.notification.IPushNotification
 import com.bugsnag.android.Bugsnag
 import expo.modules.ApplicationLifecycleDispatcher
 import chat.rocket.reactnative.networking.SSLPinningPackage;
+import chat.rocket.reactnative.notification.CustomPushNotification;
 import com.reactnativecommunity.viewpager.RNCViewPagerPackage;
 
 open class MainApplication : Application(), ReactApplication, INotificationsApplication {
