+++ android/app/src/main/java/chat/rocket/reactnative/notification/Callback.java
@@ -1,4 +1,4 @@
-package chat.rocket.reactnative;
+package chat.rocket.reactnative.notification;
 
 import android.os.Bundle;
